pipes: move struct pipe_info declaration to pipes.h
AutoFS will need to create write pipe end file descriptor, if it was closed.
Thus, pipe_info structure have to be exported.
