+++ pipes.c
@@ -15,23 +15,6 @@
 #include "protobuf/pipe.pb-c.h"
 #include "protobuf/pipe-data.pb-c.h"
 
-/*
- * The sequence of objects which should be restored:
- * pipe -> files struct-s -> fd-s.
- * pipe_entry describes  pipe's file structs-s.
- * A pipe doesn't have own properties, so it has no object.
- */
-
-struct pipe_info {
-	PipeEntry		*pe;
-	struct list_head	pipe_list;	/* All pipe_info with the same pipe_id
-						 * This is pure circular list without head */
-	struct list_head	list;		/* list head for fdinfo_list_entry-s */
-	struct file_desc	d;
-	unsigned int		create : 1,
-				reopen : 1;
-};
-
 static LIST_HEAD(pipes);
 
 static void show_saved_pipe_fds(struct pipe_info *pi)
